Placed the "New Trial" menu button into a "File" menu, which now also contains "Open Trial" and "Save Trial" routines, including keyboard shortcuts (Ctrl+N,O,S, respectively). The word choice of the new and load trial dialogs has been changed to be more clear.
Spline filtering (in particular, the down-sampled list used during the filtering process) has been changed to force inclusion of the last point of a trajectory. Hence, the smoothed per camera path must include the first and last points of a tracking sequence. Previously, filter lengths that led to smoothing that did not include the last point could yield unreasonably deviant bat coordinates near the end of the trial.